+++ lib/ui/events2/Event2SearchPanel.dart
@@ -182,6 +182,8 @@ class _Event2SearchPanelState extends State<Event2SearchPanel> implements Notifi
 
   @override
   Widget build(BuildContext context) {
+    // TBD: Replace with PopScope
+    // ignore: deprecated_member_use
     return WillPopScope(onWillPop: () => AppPopScope.back(_onHeaderBarBack), child: Platform.isIOS ?
       BackGestureDetector(onBack: _onHeaderBarBack, child:
         _buildScaffoldContent(),